+++ sshplex/__init__.py
@@ -1,4 +1,4 @@
 """SSHplex - SSH Connection Multiplexer"""
-__version__ = "1.6.0"
+__version__ = "1.6.1"
 __author__ = "MJAHED Sabri"
 __email__ = "contact@sabrimjahed.com"